# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ISS Task Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      summary: List holidays
      responses:
        "200":
          description: Holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create holiday (Boss only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete holiday (Boss only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete (clear) a single notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications read for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAllReadInput"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /notifications/clear-all:
    post:
      operationId: clearAllNotifications
      tags: [notifications]
      summary: Delete (clear) all notifications for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAllReadInput"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
components:
  schemas:
...
    Holiday:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        name:
          type: string
        day:
          type: string
        type:
          type: string
      required: [id, date, name, day, type]
    HolidayInput:
      type: object
      properties:
        date:
          type: string
        name:
          type: string
        day:
          type: string
        type:
          type: string
      required: [date, name, day, type]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        message:
          type: string
        taskId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, type, message, read, createdAt]
    MarkAllReadInput:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
...
    SuccessResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]